backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arxiv>=2.3.1",
    "fastapi[standard]>=0.128.0",
    "google-genai>=1.56.0",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "llama-index>=0.14.10",
    "llama-index-core>=0.14.10",
    "llama-index-embeddings-huggingface>=0.6.1",
    "llama-index-llms-ollama>=0.9.1",
    "llama-index-readers-file>=0.5.6",
    "llama-index-vector-stores-qdrant>=0.9.0",
    "openai>=2.14.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "redis>=7.1.0",
    "sentence-transformers>=5.2.0",
    "sqlalchemy>=2.0.45",
]
